backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
yt-dlp==2025.1.12
openai==1.51.2
faiss-cpu>=1.8.0
numpy==1.26.4

